fix(engine): silence clang -Wconversion errors in data_dict.cpp
le16() and the \uXXXX hex loop tripped -Werror,-Wimplicit-int-conversion
/ -Wsign-conversion on the ninja-clang CI, breaking the build on main for
every open PR. Wrap le16's OR in an explicit uint16_t cast and index the
escape buffer with size_t. Mirrors PR #78 (whose dead-trim hunk is moot —
that function was already removed).

+++ src/engine/data_dict.cpp
@@ -92,8 +92,9 @@ static inline std::uint32_t le32(const std::string& buf, std::size_t off) {
 }
 
 static inline std::uint16_t le16(const std::string& buf, std::size_t off) {
-    return  static_cast<std::uint16_t>(static_cast<std::uint8_t>(buf[off]))
-          | (static_cast<std::uint16_t>(static_cast<std::uint8_t>(buf[off+1])) << 8);
+    return static_cast<std::uint16_t>(
+              static_cast<std::uint16_t>(static_cast<std::uint8_t>(buf[off]))
+            | (static_cast<std::uint16_t>(static_cast<std::uint8_t>(buf[off+1])) << 8));
 }
 
 // Split a string on embedded NUL bytes.
@@ -201,7 +202,7 @@ static std::unordered_map<std::string,std::string> json_parse_flat(const std::st
                             unsigned cp = 0;
                             bool ok = true;
                             for (int k = 1; k <= 4; ++k) {
-                                char h = s[i + k];
+                                char h = s[i + static_cast<std::size_t>(k)];
                                 unsigned nib = 0;
                                 if      (h >= '0' && h <= '9') nib = static_cast<unsigned>(h - '0');
                                 else if (h >= 'a' && h <= 'f') nib = static_cast<unsigned>(h - 'a') + 10u;
